Almost done!
Now the PHP generates a fancy schmancy
garish-art-style-but-excellent-formatting HTML page, with an embed at
the top and an image at the bottom.

Next up: The issue page and support for inline images! Mostly the issue
page though, to get this out before the career fair. After that, tag
support and a better way to upload stuff - I’m thinking chrome
extension?